requirements.txt
asgiref==3.5.1
attrs==21.4.0
Django==4.0.4
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
drf-spectacular==0.22.1
inflection==0.5.1
jsonschema==4.4.0
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
